requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
flask>=2.0.0
Pillow>=8.0.0
aiohttp>=3.8.0
requests>=2.26.0 